Fix HOPO calculations in AutoGH
There were two errors in the HOPO calcuation made in AutoGH.

The first is that the code assumed that both a HOPO and the note that
preceded it cannot be chords.  That is incorrect: While the HOPO'ed
note cannot be a chord, the preceding note can be a chord as long as
it doesn't contain the HOPO'ed note.

Second, the time threshold for whether a note can be a HOPO or not
is incorrect for RockBand games.

In RockBand games, the time threshold for considering whether a note
is a HOPO can vary. By default, it is 170 ticks but it can be more or
less.  This data is recorded in a config file in the game but is not
stored in the MIDI file.

This change allows the threshold for games and an override for
individual songs to be stored in RhythmGames.xml. The songs are indexed
by title, which is stored as the name of track 0 in RockBand tracks.

If not specified, it continues to use the old formula of a 12th note
(which seems to correspond to 160 for most songs).

This was tested with Black Hole Sun (on Rock Band) and by getting the
100% with HOPO achievement in Beatles Rock Band.

Information on HOPO thresholds came from
https://github.com/ajanta/rock-band-chart-generator

The songs added to the config are only the songs I can validate (the
on-disk Rock Band, Lego Rock Band, and Rock Band Beatles).

This was validated by playing through all three Beatles Rock Band songs
with achievements for 100% HOPO as well as playing through Black Hole
Sun (in Rock Band) which is one of the songs with a non-standard
threshold that failed badly with the previous settings.